Merging refactors

+++ ExamBackend/ExamBackend/src/main/java/com/company/ExamBackend/service/impl/ExamCandidateServiceImpl.java
@@ -2,12 +2,19 @@ package com.company.ExamBackend.service.impl;
 
 import com.company.ExamBackend.dto.CandidateResponseDTO;
 import com.company.ExamBackend.dto.CandidateDashboardDTO;
+import com.company.ExamBackend.exception.ExamCandidateNotFoundException;
+import com.company.ExamBackend.exception.InvalidActionException;
 import com.company.ExamBackend.mapper.CandidateMapper;
 import com.company.ExamBackend.model.ExamCandidate;
 import com.company.ExamBackend.repository.ExamCandidateRepo;
 import com.company.ExamBackend.service.ExamCandidateService;
 import lombok.AllArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.Duration;
 import java.time.Instant;
@@ -19,45 +26,71 @@ public class ExamCandidateServiceImpl implements ExamCandidateService {
 
     private final ExamCandidateRepo examCandidateRepo;
 
+    private final CandidateMapper candidateMapper;
+
     //Lists all candidates for respective exam.
     @Override
-    public List<CandidateResponseDTO> getCandidates(String examId) {
-        List<ExamCandidate> candidates = examCandidateRepo.findByExamId(examId);
-        return CandidateMapper.toDTOList(candidates);
+    public Page<CandidateResponseDTO> getCandidates(
+            String examId,
+            String adminEmail,
+            int page,
+            int size) {
+        Sort sort = Sort.by(Sort.Direction.DESC, "id");
+        Pageable pageable = PageRequest.of(page, size, sort);
+        Page<ExamCandidate> candidates = examCandidateRepo.findByExamIdAndAdminEmail(examId,adminEmail,pageable);
+        return candidateMapper.toDTOList(candidates);
     }
 
-    //Removes candidate from exam, to be implemented later.
+    //Removes candidate from exam
+    @Transactional
     @Override
     public void removeCandidate(String examId, String email) {
+        ExamCandidate reqCandidate = examCandidateRepo.findByExamIdAndEmail(examId, email);
+        if(reqCandidate == null)
+        {
+            throw new ExamCandidateNotFoundException("ExamCandidate of email "+email+" not found");
+        }
+
+        if(!checkDeletionValidity(reqCandidate)) {
+            throw new InvalidActionException("Candidate of status "+reqCandidate.getStatus()+" cannot be deleted.");
+        }
+
         examCandidateRepo.deleteByExamIdAndEmail(examId, email);
     }
 
     //Sends exam details to candidate dashboard.
     @Override
     public List<CandidateDashboardDTO> getCandidateDashboard(String email) {
-        Instant now = Instant.now();
-        return examCandidateRepo.findByEmail(email).stream()
-                .filter(examCandidate -> "INVITED".equals(examCandidate.getStatus()))
-                .filter(candidate -> {
-                    Instant start = candidate.getExam().getStartTime();
-                    Instant end = candidate.getExam().getEndTime();
-                    int durationMins = candidate.getExam().getDuration();
-                    //Checks if the exam has started and hasn't ended yet
-                    boolean isWithinWindow = (now.equals(start) || now.isAfter(start)) && now.isBefore(end);
-                    if (!isWithinWindow) return false;
-                    //Checks if there is enough time left to actually complete the exam
-                    Instant latestPossibleFinishTime = now.plus(Duration.ofMinutes(durationMins));
-
-                    return !latestPossibleFinishTime.isAfter(end);
-                })
-                .map(candidate -> CandidateDashboardDTO.builder()
-                        .examId(candidate.getExam().getId())
-                        .title(candidate.getExam().getTitle())
-                        .duration(candidate.getExam().getDuration())
-                        .startTime(candidate.getExam().getStartTime())
-                        .endTime(candidate.getExam().getEndTime())
-                        .candidateStatus(candidate.getStatus())
-                        .build())
+        return examCandidateRepo.findActiveDashboardExams(email, Instant.now())
+                .stream()
+                .map(candidateMapper::toDashboardDTO)
                 .toList();
+    }
+
+    // ======================================================================================
+    // Helper methods
+    // ======================================================================================
+
+    private boolean isWithinValidTimeWindow(ExamCandidate candidate, Instant now) {
+        Instant start = candidate.getExam().getStartTime();
+        Instant end = candidate.getExam().getEndTime();
+        int durationMins = candidate.getExam().getDuration();
+
+        boolean isActive = (now.equals(start) || now.isAfter(start)) && now.isBefore(end);
+        if (!isActive) return false;
+
+        Instant latestPossibleFinish = now.plus(Duration.ofMinutes(durationMins));
+        return !latestPossibleFinish.isAfter(end);
+    }
+
+    private boolean checkDeletionValidity(ExamCandidate candidate) {
+        return (
+                candidate.getStatus().equals("INVITED")
+                || candidate.getStatus().equals("UNINVITED")
+        );
+    }
+
+    private ExamCandidate getExamCandidate(String examId, String email) {
+        return examCandidateRepo.findByExamIdAndEmail(examId, email);
     }
 }
